+++ Properties/AssemblyInfo.cs
@@ -5,8 +5,8 @@ using System.Runtime.Versioning;
 
 [assembly: AssemblyCompany("NiumaClaw.Team")]
 [assembly: AssemblyConfiguration("Release")]
-[assembly: AssemblyFileVersion("1.0.4.0")]
-[assembly: AssemblyInformationalVersion("1.0.4+mac-dmg-client.20260701")]
+[assembly: AssemblyFileVersion("1.0.5.0")]
+[assembly: AssemblyInformationalVersion("1.0.5+mac-dmg-client-url.20260701")]
 [assembly: AssemblyProduct("NiumaClaw.Team")]
 [assembly: AssemblyTitle("NiumaClaw.Team")]
-[assembly: AssemblyVersion("1.0.4.0")]
+[assembly: AssemblyVersion("1.0.5.0")]